requirements.txt
PyYAML
coloredlogs
pyqt5-tools
pure-python-adb